+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ## Compatibility Notes
 
--
+- [core] Tightened the return type of `getAuthHeaders()` and `buildAuthorizationHeaders()` methods.
 
 ## New Functionality
 
@@ -27,6 +27,8 @@
 ## Fixed Issues
 
 - [odata] Fix parsing of `Edm.Time` missing some parts of the time e.g. `PT1H21S`.
+- [util] `setGlobalLogLevel(level)` sets log levels of all existing loggers globally.
+- [util] Fix `equal(item1, item2)` method to return `false` when either object is `null` or `undefined` 
 
 
 # 1.50.0
@@ -42,10 +44,13 @@ Blog: TBD<br>
 ## Improvements
 
 - [generator] Include and exclude options added to the `transpileDirectory()` method
+- [openapi-generator] Replace `oclif` library with `yargs`. This is a refactoring without functional changes.
 
 ## Fixed Issues
 
 - [core] Fix a type error, when `moment()` is used in filtering an OData field of type `Edm.DateTimeOffset`
+- [core] Fix missing `Proxy-Authorization` header if custom authorization headers are set.
+- [core] Remove wrongly given warning that custom headers are given and will overwrite destination headers.
 
 # 1.49.0